+++ os/api/src/commonMain/kotlin/com/wasmo/api/routes/RouteCodec.kt
@@ -1,5 +1,7 @@
 package com.wasmo.api.routes
 
+import com.wasmo.framework.Url
+
 interface RouteCodec {
   fun decode(url: Url): Route
   fun encode(route: Route): Url
