+++ os/framework/src/commonMain/kotlin/com/wasmo/framework/Url.kt
@@ -1,4 +1,4 @@
-package com.wasmo.api.routes
+package com.wasmo.framework
 
 data class Url(
   val scheme: String,
