+++ os/server/accounts/real/src/jvmMain/kotlin/com/wasmo/accounts/ClientAuthenticator.kt
@@ -1,5 +1,7 @@
 package com.wasmo.accounts
 
+import com.wasmo.framework.UserAgent
+
 interface ClientAuthenticator {
   fun updateSessionCookie(): SessionCookie
   fun signOutSessionCookie()
@@ -7,25 +9,5 @@ interface ClientAuthenticator {
 
   interface Factory {
     fun create(userAgent: UserAgent): ClientAuthenticator
-  }
-
-  /** Minimal HTTP request/response API for getting and setting session cookies. */
-  interface UserAgent {
-    val userAgent: String?
-
-    val ip: String
-
-    /** Returns the cookie from the request, or set on the response if that exists. */
-    fun getCookie(name: String): String?
-
-    fun setCookie(
-      name: String,
-      value: String,
-      secure: Boolean,
-      httpOnly: Boolean,
-      domain: String,
-      path: String,
-      maxAgeSeconds: Long,
-    )
   }
 }
