+++ os/server/accounts/real/src/jvmMain/kotlin/com/wasmo/accounts/RealClientAuthenticator.kt
@@ -1,8 +1,8 @@
 package com.wasmo.accounts
 
-import com.wasmo.accounts.ClientAuthenticator.UserAgent
 import com.wasmo.deployment.Deployment
 import com.wasmo.framework.UnauthorizedUserException
+import com.wasmo.framework.UserAgent
 import com.wasmo.support.tokens.newToken
 import dev.zacsweers.metro.Assisted
 import dev.zacsweers.metro.AssistedFactory
