+++ os/server/accounts/real/src/jvmMain/kotlin/com/wasmo/accounts/invite/CreateInviteRpc.kt
@@ -5,9 +5,9 @@ import com.wasmo.accounts.Client
 import com.wasmo.api.CreateInviteRequest
 import com.wasmo.api.CreateInviteResponse
 import com.wasmo.api.routes.InviteRoute
-import com.wasmo.api.routes.toHttpUrl
 import com.wasmo.calls.CallDataService
 import com.wasmo.framework.Response
+import com.wasmo.framework.toHttpUrl
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.SingleIn
 import wasmo.sql.SqlDatabase
