+++ os/server/ktor/src/jvmMain/kotlin/com/wasmo/ktor/HttpActionBinder.kt
@@ -1,10 +1,10 @@
 package com.wasmo.ktor
 
-import com.wasmo.accounts.ClientAuthenticator.UserAgent
-import com.wasmo.api.routes.Url
 import com.wasmo.framework.Request
 import com.wasmo.framework.Response
 import com.wasmo.framework.ResponseBody
+import com.wasmo.framework.Url
+import com.wasmo.framework.UserAgent
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.serializer
