+++ os/server/ktor/src/jvmMain/kotlin/com/wasmo/ktor/HttpActionSource.kt
@@ -1,0 +1,6 @@
+package com.wasmo.ktor
+
+interface HttpActionSource {
+  context(binder: HttpActionBinder)
+  fun bindActions()
+}
